E4S: Car/Landslide Uptime trigger improvements
Changed the way the trigger works so that isn't a need for user input in selecting/changing names with roles etc. when/if party composition switches or if we sub people out.

The trigger now checks for yellow markers during Car/Landslide and identifies which DPS has the marker, it checks if it is a melee job (MNK, DRG, NIN, SAM) or a ranged job (BRD, MCH, DNC, BLM, SMN, RDM) and will make the proper callout.