tea: updated clone capture regex for Fate Calibration
Debuff icons seem to be appearing by dry firing the trigger. I have two possible theories at the moment:

First
-----
Since the images are handled by a seperate trigger, the x and y coordinates aren't properly captured and the image actually appears in the TOP-LEFT corner of your screen instead of inside the Fate Calibration box.

Second
------
Clone capture isn't working as intended/properly. This may very well be the issue here (hence the updated regex) as the TTS callout for the debuffs does not fire as well. I might have a fix in the future if the updated regex actually works but callouts/images still aren't happening.