tea: unfucked Nisi (among other things)
Nisi
====
- [repo] This was accidentally deleted some time ago and has been re-added. Sorry about that lol...
- [repo/sa] Final (4th) pass before Gavel has been finicky. Added new conditions to see if it actually fires without any hiccups.

Final Word
==========
- [repo] Debuff images has been added for both Light and Dark (Marks and Beacons).

Note: Strike Spark from E6S is somewhat broken and occasionally calls out the wrong safe zone. But since we're not clearing Savage anymore (and even if we are, we're eating one ball north), this remains a low priority fix and won't happen anytime soon (or ever).