trials: updated final trial triggers
Triggers for this fight should be working (hopefully) as of now and are zone locked.

Triggers that assist in tracking abilities such as Imbued Saber and Quintuplecast have not been tested and may or may nnot function properly. So, please open your eyes and see the boss's casts and animation (as you always should be doing even with working triggers!).