alliance: added initial triggers for The Puppets' Bunker
If all triggers fail to go off, the zone regex is wrong. Copy the zone folder (02 - The Puppets' Bunker) to your local triggers folder and turn the zone requirement off manually.

(How to) Right Click on "02 - The Puppets' Bunker" select "Edit" and uncheck the only checked box that's under the "General" tab.