bozja: updated Peerifool debuff tracking reset
- Reset for debuff tracking will now also trigger on player death.

(The only reason for this trigger to break is if you somehow time out on this CE, which I think is almost impossible.)

Sidenote: One option for not allowing triggers to go off if you are not in a CE is to add a "in combat" checker. However, if you are fighting while nearby an ongoing CE, the triggers will still fire. So for the moment, this will not be added as a fix.