bozja: added Castrum Lacus Litore
Current issues are that during split fights (at the gates and against Lyon & Dawon) there are no way to split the triggers into separate encounters. Meaning that (for example) triggers for Dawon will go off when you are fighting Lyon. The logical would be to add a checker, but that is easier said than done. Although it is on my to-do list.