bozja: added 0x21 lines as CE checker
All Critical Engagements and Duels will now use 0x21 lines to determine if the player is actually inside. This will prevent triggers from going off if you are a bystander.

(The Battle for Castrum Lacus Litore does not use the 0x21 lines as checker. Although I may add it some time in the future is for consistency sake. As of now, Castrum triggers are not fired when players are in the field so this change isn't a priority.)