e10s: fixes
[ Dualspell ]
=============
Dualspell uses an offset like Limit Cut in TEA. This affects the ID for both the 1-2-3 headmarkers and the stack marker. For now, I've applied a fix to show the 1-2-3 headmarkers.

[ Umbra Smash ]
================
Added a reset for Umbra Smash after tank cleave or stack. The reset wasn't there and would be giving tanks the wrong callout after the first Umbra Smash.

[ Swath of Silence ]
====================
Changed the callout from "Under Boss" to "Clone Cleave" because of Pitch Bog 2.