e12s: text aura fixes
[ Text Auras ]
==============
The following text auras should now be working as intended:
- Short/Long Fire/Blizzard during Basic
- Debuff string during Intermediate
- Double Wind and Shadoweye during Advanced

This was a hindsight on my end as I assumed the list variable I was using as a checker worked as it said it did. (It did not, hence the checker would always be one variable short.) I've made some crude adjustments and tested them in-game to make sure they worked this time (especially the debuff string for Intermediate). The TTS trigger for Intermediate will remain as a fallback just in case text fails again for whatever reason.

NOTE: Prismatic Deception trigger for E11S might stilll be somewhat broken. As a rule of thumb, always do the mechanic as intended by your group instead of fully relying on triggers (especially if it has many moving parts behind the scenes) to do the mechanic for you.