dr: minor (and final) savage support (see desc)
- Added minor raidwide/buster callouts in savage for some of the existing bosses in normal.
- Added basic support for Stygimoloch Lord.
(Note: This is the boss fight before the Queen. The duel is against Stygimoloch Warrior, not Lord.)

This will probably be the only support that DRS will be getting from my end. Reason being that Savage will be something that will have an even smaller number of clears compared to Ultimate due to how matchmaking is handled for this content. Also, certain mechanics are handled completely different from normal mode (such as the Sword/Shield Omen).

tl;dr: DRS triggers are too much of a hassle, a big headache, and I'm lazy. (I'm also not Rank 15 and can't even get half a group of 8, let alone 47 other people to even try to attempt Savage.)