ucob: added countdown for Fireball tether
Fireball phase will be THE MOST BUSIEST on-screen phase with countdown text auras for Fireball tethers and Thunderstruck, Nael's RP, Doom counter and Doom cleanse counter. For now, I'll be leaving it as it is but I would like to reduce the screen clutter going forward.

Also re-enabled E11S zone lock on the folder. (Which I forgot to do after pushing the final fix for multi-tethers.)