fixed: added Crystallize reset to Equinox
Hydaelyn gains a status effect for Anthelion (donut/dynamo/in mechanic) and Highest Holy (out/chariot mechanic) in logs. This line is missing when she casts Equinox (+ line/intercardinal mechanic). These log lines are used to reset the Crystallize tracker but with Equinox missing, the tracker doesn't reset when this mechanic is executed and causes some minor tracking errors in the fight.