raids (6.0): updated backend variable stuff
- Removed all periods on variable names as to not cause issues down the line.

One such area where this was an issue was in P2S during Mark of the Tides tracking. When referencing a list variable, any periods in the name of the list variable will cause the entire thing to not function. The initial fix was to replace periods with underscores, but I've ultimately decided on just removing them for all future triggers.

Some older triggers of past content still use this naming convention (with the period) and I don't really want to scrub through all of them. Those triggers still work and (hopefully) won't cause any issues down the line. For now, those will remain untouched.