dsw: added doorboss triggers
[ Offset ]
==========
Offset checker will happen twice in the fight: during the door boss and during Thordan. Usually, just having it check once would be good enough. However, in the event that someone restarts ACT after clearing the door boss, the offset will no longer be saved. This is just to make sure that Triggernometry checks for the offset again and apply it.