[v6.21.221018a] added CDH soundbytes
[ Limit Break ]
===============
- Updated GNB's Gunmetal Soul to make "MY TURN" appear 50% more often over the others.

[ CDH ]
=======
Abilities from specific job skills will trigger a soundbyte when it is a Critical Direct Hit.
- Afflatus Misery (WHM)
- Blast Arrow (BRD)
- Communio (RPR)
- Confiteor (PLD)
- Double Down (GNB)
- Hissatsu: Senei (SAM)
- Hyosho Ranryu (NIN)
- Phantom Rush (MNK)
- Primal Rend (WAR)
-- Disabled by default as this is always a CDH.
- Resolution (RDM)
- Shadowbringer (DRK)
-- Only triggers for self. Esteem doesn't trigger the soundbyte.
- Stardiver (DRG)
- Technical Finish (DNC)
-- Only applies to Quad finishes. Single, Double and Triple will not trigger anything.
- Xenoglossy (BLM)

SCH and AST have no hard hitting skills and thus, not included. SGE's Phlegma III is currently being considered. MCH's tools are also being considered or probably replaced with Wildfire or Queen's Crowned Collider. SMN is on hold due to Akh Morn and Revelation not originating from the caster and might cause issues if there are multiple SMNs in the party.